+++ base/sparse/umfpack.jl
@@ -11,8 +11,8 @@ using ..SparseArrays
 import ..SparseArrays: increment, increment!, decrement, decrement!, nnz
 
 include("umfpack_h.jl")
-mutable struct MatrixIllConditionedException <: Exception
-    message::AbstractString
+struct MatrixIllConditionedException <: Exception
+    msg::AbstractString
 end
 
 function umferror(status::Integer)